add new sign
